Fix: Skip relationship one to many and many to one

+++ src/main/java/ar/com/franciscoderamo/portfolio/interfaces/ExperienceInterface.java
@@ -9,8 +9,8 @@ public interface ExperienceInterface {
     
     public void saveExperience(ExperienceModel experience);
     
-    public void deleteExperience (Long id);
+    public void deleteExperience (int id);
     
-    public ExperienceModel findExperience(Long id);
+    public ExperienceModel findExperience(int id);
 
 }
